refactor: rename voice -> live
